Bump version to 2.0: Changed scheduler to HFSC and export flow labels
The HTB scheduler has been in the Linux kernel for lots of years, but a
better scheduler has been developed, called HFSC. It has a much better
link-sharing algorithm which we use here. It also has support for real-time
(hard) latency guarantees, but I couldn't find any good documentation that
explains how to exactly convert required bitrate and max latency numbers
applications give in their documentation to proper "rt" service curve
configuration parameters. But using just "ls" service curves still
impressively outperforms the older HTB scheduler.

The flow definition labels are now exported into a CSV file so that you can
use them in other scripts without having to copy-n-paste them to that other
software. This should avoid the flow definitions getting out of sync with
other monitoring software.

The flow definitions are printed on startup along with their rates and
percentages. The sum of all the percent values are also printed so it is
easy to see if they sum up to 100%.

The stop_me function didn't actually check the qdisc on the defined
interface, but used the example value (ppp0). This should now be fixed.

Updated changelog and README.